Object detection + restructure + depth_object
New package created for the detection of objects and integrating depth into that object detection. Just the initial package is created and not the final implementation.

Code is a little bit restructured. Earlier there was a package for depth only, but now the object and the depth package is combined and a combined package is created for implementation.