Structure change for the depth and object module
The structure of the interfaced code is changed now and the object and depth module has oops support now. The main middle zed object has to be shared by every feature.

More updates to this upgrade coming soon and the object module has to be handled in the same way.